server/requirements.txt
pyserial>=3.5
websockets>=12.0
numpy>=1.24
scikit-learn>=1.3
flask>=3.0
torch>=2.0
filterpy>=1.4
